requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
Django==3.0.8
django-cors-headers==3.4.0
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-python3-ldap==0.11.3
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
flower==0.9.5
future==0.18.2
humanize==2.5.0
importlib-metadata==1.7.0
kombu==4.6.11
ldap3==2.7
prometheus-client==0.8.0
pyasn1==0.4.8
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
tornado==6.0.4
vine==1.3.0
zipp==3.1.0
